// src/content/generated/lessonsIndex.ts
export const weeks: WeekInfo[] = [
  {
    "week": 1,
    "title": "Grundlagen",
    "lessonIds": [
      1,
      2,
      3,
      4,
      5,
      6,
      7
    ]
  },
  {
    "week": 2,
    "title": "Diagnostik und Behandlungen",
    "lessonIds": [
      8,
      9,
      10,
      11,
      12,
      13,
      14
    ]
  },
  {
    "week": 3,
    "title": "Spezielle Themen und Praxisalltag",
    "lessonIds": [
      15,
      16,
      17,
      18,
      19,
      20,
      21
    ]
  }
];